Maid init.lua
Keep a history of items to be cleaned later. Returns insert and delete functions that can be chained, delete can be called automatically by providing a signal as first argument or manually by simply calling `delete()`

It's designed to keep code clean by moving it all the way to the bottom. Might not suit all use-cases which is why it can be used like any other maid too.